refactor: react-table and data-grid-table
