--- examples/table/mui/advancedTable/src/pages/dataGrid/index.ts
@@ -1,1 +1,0 @@
-export * from "./basicDataGrid";
